Crate paralegal_flow

Ties together the crate and defines command line options.

While this is technically a “library”, it only is so for the purposes of being able to reference the same code in the two executables paralegal_flow and cargo-paralegal-flow (a structure suggested by rustc_plugin).

Re-exports

pub extern crate either;

pub extern crate rustc_index;

pub use crate::ann::db::MarkerCtx;

pub use paralegal_spdg as desc;

Modules

ana

Data-and control flow analyzer and inliner.

ann

args 🔒

Command line arguments and parsing.

consts

Constants used globally.

dbg

Helpers for debugging

discover 🔒

MIR visitor (CollectingVisitor) that populates the MarkerDatabase and discovers functions marked for analysis.

pdg 🔒

rust

Exposes the rustc external crates (this mod is just to tidy things up).

serializers

serde serializers, most used for JSON debugging output in crate::dbg.

utils

Utility functions, general purpose structs and extension traits

Macros

sym_vec

Conveniently create a vector of Symbols. This way you can just write sym_vec!["s1", "s2", ...] and this macro will make sure to call Symbol::intern

Structs

AnalysisCtrl

Arguments that control the flow analysis

ArgWrapper 🔒

Top level argument structure. This is only used for parsing. The actual configuration of paralegal-flow Args which is stored in args. cargo_args is forwarded and _progname is only to comply with the calling convention of cargo (it passes the program name as first argument).

Args

BuildConfig

Additional configuration for the build process/rustc

Callbacks 🔒

DepConfig

Dependency specific configuration

DfppPlugin

A struct so we can implement rustc_plugin::RustcPlugin

DumpArgs

Collection of the DumpOptions a user has set.

HashMap

A hash map implemented with quadratic probing and SIMD lookup.

HashSet

A hash set implemented as a HashMap where the value is ().

ModelCtrl

NoopCallbacks 🔒

Symbol

An interned string.

Enums

Either

The enum Either with variants Left and Right is a general purpose sum type with two cases.

Constants

CARGO_ENCODED_RUSTFLAGS

Functions

add_to_rustflags 🔒