paralegal_flow
Module ana

Data-and control flow analyzer and inliner.

Analysis starts with the construction of SPDGGenerator from a CollectingVisitor and then calling analyze.

Modules

graph_converter 🔒

inline_judge 🔒

Structs

InlineJudge

The interpretation of marker placement as it pertains to inlining and inline elision.

MyCallback 🔒

SPDGGenerator

Read-only database of information the analysis needs.

Functions

def_info_for_item 🔒

def_kind_for_item 🔒

default_index 🔒

dirty_try_resolve_func_id 🔒

path_for_item 🔒

src_loc_for_span 🔒

type_info_sanity_check 🔒

Checks the invariant that SPDGGenerator::collect_type_info should produce a map that is a superset of the types found in all the types maps on SPDG.

with_reset_level_if_target 🔒

A higher order function that increases the logging level if the target matches the one selected with the debug flag on the command line (and reset it afterward).