paralegal_flow::ana
Module graph_converter

Modules

call_string_resolver 🔒

Resolution of CallStrings to [FnResolution]s.

Structs

GraphConverter

Structure responsible for converting one DepGraph into an SPDG.

Functions

assert_edge_location_invariant 🔒

entrypoint_is_async 🔒

expect_stmt_at 🔒

Find the statement at this location or fail.

get_parent 🔒

If did is a method of an impl of a trait, then return the DefId that refers to the method on the trait definition.