Struct paralegal_flow::ana::GraphConverter

struct GraphConverter<'tcx, 'a, C> {
    generator: &'a SPDGGenerator<'tcx>,
    target: FnToAnalyze,
    dep_graph: Rc<DepGraph<'tcx>>,
    local_def_id: LocalDefId,
    known_def_ids: &'a mut C,
    types: HashMap<Node, Types>,
    index_map: Box<[Node]>,
    spdg: SPDGImpl,
}

Structure responsible for converting one DepGraph into an SPDG.

Intended usage is to call Self::new_with_flowistry to initialize, then Self::make_spdg to convert.

Fields
generator: &'a SPDGGenerator<'tcx>

The parent generator

target: FnToAnalyze

Information about the function this PDG belongs to

dep_graph: Rc<DepGraph<'tcx>>

The flowistry graph we are converting

local_def_id: LocalDefId

Same as the ID stored in self.target, but as a local def id

known_def_ids: &'a mut C

Where we write every DefId we encounter into.

types: HashMap<Node, Types>

A map of which nodes are of which (marked) type. We build this up during conversion.

index_map: Box<[Node]>

Mapping from old node indices to new node indices. Use Self::register_node to insert and Self::new_node_for to query.

spdg: SPDGImpl

The converted graph we are creating

Implementations

impl<'a, 'tcx, C: Extend<DefId>> GraphConverter<'tcx, 'a, C>

fn new_with_flowistry( generator: &'a SPDGGenerator<'tcx>, known_def_ids: &'a mut C, target: FnToAnalyze ) -> Result<Self>

Initialize a new converter by creating an initial PDG using flowistry.

fn tcx(&self) -> TyCtxt<'tcx>

fn marker_ctx(&self) -> &MarkerCtx<'tcx>

fn entrypoint_is_async(&self) -> bool

Is the top-level function (entrypoint) an async fn

fn expect_stmt_at( &self, loc: GlobalLocation ) -> Either<&'tcx Statement<'tcx>, &'tcx Terminator<'tcx>>

Find the statement at this location or fail.

fn register_node(&mut self, old: Node, new: NodeInfo) -> Node

Insert this node into the converted graph, return it’s auto-assigned id and register it as corresponding to old in the initial graph. Fails if there is already a node registered as corresponding to old.

fn new_node_for(&self, old: Node) -> Node

Get the id of the new node that was registered for this old node.

fn determine_node_kind( &mut self, weight: &DepNode<'tcx> ) -> (NodeKind, bool, Vec<Identifier>)

Try to discern if this node is a special NodeKind. Also returns if the location corresponds to a function call for an external function and any marker annotations on this node.

fn determine_place_type(&self, weight: &DepNode<'tcx>) -> PlaceTy<'tcx>

Reconstruct the type for the data this node represents.

fn annotations_for_function( &self, function: DefId, filter: impl FnMut(&MarkerAnnotation) -> bool ) -> (Vec<Identifier>, Option<DefId>)

Fetch annotations item identified by this id.

The callback is used to filter out annotations where the “refinement” doesn’t match. The idea is that the caller of this function knows whether they are looking for annotations on an argument or return of a function identified by this id or on a type and the callback should be used to enforce this.

fn handle_node_types( &mut self, i: Node, weight: &DepNode<'tcx>, is_external_call_source: bool )

Check if this node is of a marked type and register that type.

fn create_flowistry_graph( generator: &SPDGGenerator<'tcx>, local_def_id: LocalDefId ) -> Result<DepGraph<'tcx>>

Create an initial flowistry graph for the function identified by local_def_id.

fn make_spdg(self) -> SPDG

Consume the generator and compile the SPDG.

fn make_spdg_impl(&mut self) -> HashMap<Node, Vec<Identifier>>

This initializes the fields spdg and index_map and should be called first

fn type_is_marked(&self, typ: PlaceTy<'tcx>, walk: bool) -> Vec<TypeId>

Return the (sub)types of this type that are marked.

fn try_as_root(&self, at: CallString) -> Option<GlobalLocation>

Similar to CallString::is_at_root, but takes into account top-level async functions

fn determine_return(&self) -> Option<Node>

Try to find the node corresponding to the values returned from this controller.

TODO: Include mutable inputs

fn determine_arguments(&self) -> Vec<Node>

Determine the set if nodes corresponding to the inputs to the entrypoint. The order is guaranteed to be the same as the source-level function declaration.