paralegal_flow::ana
Struct SPDGGenerator

pub struct SPDGGenerator<'tcx> {
    ctx: Pctx<'tcx>,
    stats: Stats,
    pdg_constructor: MemoPdgConstructor<'tcx, u32>,
    functions_to_analyze: Vec<FnToAnalyze>,
}

Read-only database of information the analysis needs.

Self::analyze serves as the main entrypoint to SPDG generation.

Fields
ctx: Pctx<'tcx>stats: Statspdg_constructor: MemoPdgConstructor<'tcx, u32>functions_to_analyze: Vec<FnToAnalyze>
Implementations

impl<'tcx> SPDGGenerator<'tcx>

pub fn new( ctx: Pctx<'tcx>, inline_judge: InlineJudge<'tcx>, stats: Stats, functions_to_analyze: Vec<FnToAnalyze>, ) -> Self

pub fn marker_ctx(&self) -> &MarkerCtx<'tcx>

pub fn tcx(&self) -> TyCtxt<'tcx>

fn handle_target( &mut self, target: &FnToAnalyze, known_def_ids: &mut impl Extend<DefId>, ) -> Result<(Endpoint, SPDG)>

Perform the analysis for one #[paralegal_flow::analyze] annotated function and return the representation suitable for emitting into Forge.

Main work for a single target is performed by GraphConverter.

pub fn analyze(&mut self) -> Result<(ProgramDescription, AnalyzerStats)>

Main analysis driver. Essentially just calls Self::handle_target once for every function in self.functions_to_analyze after doing some other setup necessary for the flow graph creation.

Should only be called after the visit.

fn make_program_description( &self, controllers: HashMap<Endpoint, SPDG>, known_def_ids: HashSet<DefId>, _targets: &[FnToAnalyze], ) -> (ProgramDescription, AnalyzerStats)

Given the PDGs and a record of all DefIds we’ve seen, compile auxillary information the policies will need into the artifact to be emitted.

fn collect_stats_and_analyzed_spans(&self) -> (AnalyzerStats, AnalyzedSpans)

fn get_prior_stats(&self) -> DumpStats

fn collect_instruction_info( &self, controllers: &HashMap<Endpoint, SPDG>, known_def_ids: &mut impl Extend<DefId>, ) -> HashMap<GlobalLocation, InstructionInfo>

Create an InstructionInfo record for each GlobalLocation mentioned in the controllers.

fn collect_type_info(&self) -> TypeInfoMap

Create a TypeDescription record for each marked type that as mentioned in the PDG.