paralegal_flow::ann
Module db

Central repository for information about markers (and annotations).

The database (MarkerDatabase) is initialized with (init) and populated with local markers by CollectingVisitor by calls to (retrieve_local_annotations_for). Then it’s transformed into a read-only MarkerCtx via From::from. The MarkerCtx is a cheap pointer to the database and responsible for answering queries about markers as the main analysis runs.

All interactions happen through the central database object: MarkerCtx.

Structs

ExternalAnnotationEntry 🔒

FunctionMarkerStat

MarkerCtx

The marker context is a database which can be queried as to whether functions or types carry markers, whether markers are reachable in bodies, etc.

MarkerDatabase

The structure inside of MarkerCtx.

Enums

MaybeMonomorphized

Either we have an Instance or a DefId if we weren’t able to resolve the generics.

Functions

check_format 🔒

load_annotations 🔒

marker_stats_as_json 🔒

parse_external_marker_file 🔒

resolve_external_markers 🔒

Given the TOML of external annotations we have parsed, resolve the paths (keys of the map) to DefIds.

Type Aliases

ExternalMarkers 🔒

RawExternalMarkers 🔒

TypeMarkerElem

TypeMarkers