paralegal_flow::ann::parse
Function delimited

pub fn delimited<'a, A, P: Parser<I<'a>, A, ErrorTree<I<'a>>> + 'a>(
    p: P,
    delim: Delimiter,
) -> impl FnMut(I<'a>) -> IResult<I<'a>, A, ErrorTree<I<'a>>>

Parse a TokenTree::Delimited with the delimiter character delim, applying the subparser p to the tokens in between the delimiters and return the result of the subparser.