paralegal_flow::ann
Struct Markers

struct Markers {
    label_marker: Vec<Symbol>,
    marker_marker: Vec<Symbol>,
    otype_marker: Vec<Symbol>,
    exception_marker: Vec<Symbol>,
}

These are pseudo-constants that are used to match annotations. In theory they never change but they use [Symbol] inside, which is only valid so long as the rustc SESSION_GLOBALS are live so we need to calculate them afresh in new.

Fields
label_marker: Vec<Symbol>

This will match the annotation #[paralegal_flow::label(...)] when using MetaItemMatch::match_extract

marker_marker: Vec<Symbol>

This will match the annotation #[paralegal_flow::marker(...)] when using MetaItemMatch::match_extract

otype_marker: Vec<Symbol>

This will match the annotation #[paralegal_flow::output_types(...)] when using MetaItemMatch::match_extract

exception_marker: Vec<Symbol>

This will match the annotation #[paralegal_flow::exception(...)] when using MetaItemMatch::match_extract