Module paralegal_flow::args

Command line arguments and parsing.

This module deliberately hides all of the struct fields we have here. The reason is that the fields and their types are used by clap as the names of the command line arguments. But we might want to change the names or their interpretation (e.g. go from a default positive argument like --strict to a default negative one e.g. --relax). That is why instead we expose methods, which do not correspond directly to those arguments but allow us to change the name and default value of the argument without having to migrate the code using that argument.

Structs

AnalysisCtrl

Args

BuildConfig

Additional configuration for the build process/rustc

ClapAnalysisCtrl 🔒

Arguments that control the flow analysis

ClapArgs

Arguments as exposed on the command line.

DUMP_ARGS_OPTIONS 🔒

DepConfig

Dependency specific configuration

DumpArgs

Collection of the DumpOptions a user has set.

MarkerControl

ParseableDumpArgs

Enums

Debugger

DumpOption 🔒

InliningDepth

LogLevelConfig

How a specific logging level was configured. (currently only used for the --debug level)

Stub

VarError 🔒

Constants

_IMPL_NUM_FromPrimitive_FOR_DumpOption 🔒

Functions

config_hash_for_file 🔒

env_var_expect_unicode 🔒

A thin wrapper around std::env::var that returns None if the variable is not present.