paralegal_flow::args
Struct Args

pub struct Args {
    verbosity: LevelFilter,
    log_level_config: LogLevelConfig,
    result_path: PathBuf,
    relaxed: bool,
    target: Option<String>,
    abort_after_analysis: bool,
    attach_to_debugger: Option<Debugger>,
    marker_control: MarkerControl,
    anactrl: AnalysisCtrl,
    dump: DumpArgs,
    build_config: (Option<PathBuf>, BuildConfig),
    cargo_args: Vec<String>,
}

Fields
verbosity: LevelFilter

Print additional logging output (up to the “info” level)

log_level_config: LogLevelConfigresult_path: PathBuf

Where to write the resulting forge code to (defaults to analysis_result.frg)

relaxed: bool

Emit warnings instead of aborting the analysis on sanity checks

target: Option<String>

Target a specific package

abort_after_analysis: bool

Abort the compilation after finishing the analysis

attach_to_debugger: Option<Debugger>

Make the compiler attach to a debugger

marker_control: MarkerControl

Additional arguments on marker assignment and discovery

anactrl: AnalysisCtrl

Additional arguments that control the flow analysis specifically

dump: DumpArgs

Additional arguments that control debug output specifically

build_config: (Option<PathBuf>, BuildConfig)

Additional configuration for the build process/rustc

cargo_args: Vec<String>

Additional options for cargo

Implementations

impl Args

pub fn target(&self) -> Option<&str>

pub fn direct_debug(&self) -> &LogLevelConfig

Returns the configuration specified for the --debug option

pub fn dbg(&self) -> &DumpArgs

Access the debug arguments

pub fn anactrl(&self) -> &AnalysisCtrl

Access the argument controlling the analysis

pub fn result_path(&self) -> &Path

the file to write results to

pub fn verbosity(&self) -> LevelFilter

Should we output additional log messages (level info)

pub fn relaxed(&self) -> bool

Warn instead of crashing the program in case of non-fatal errors

pub fn abort_after_analysis(&self) -> bool

pub fn build_config(&self) -> &BuildConfig

pub fn hash_config(&self, hasher: &mut impl Hasher)

pub fn marker_control(&self) -> &MarkerControl

pub fn cargo_args(&self) -> &[String]

pub fn attach_to_debugger(&self) -> Option<Debugger>

pub fn setup_logging(&self)