paralegal_flow::args
Struct ClapArgs

pub struct ClapArgs {

    verbose: bool,
    debug: bool,
    trace: bool,
    debug_target: Option<String>,
    result_path: PathBuf,
    relaxed: bool,
    strict: bool,
    target: Option<String>,
    abort_after_analysis: bool,
    attach_to_debugger: Option<Debugger>,
    anactrl: ClapAnalysisCtrl,
    marker_control: MarkerControl,
    dump: ParseableDumpArgs,
    cargo_args: Vec<String>,

}

Arguments as exposed on the command line.

You should then use try_into to convert this to [Args], the argument structure used internally.

Fields
verbose: bool

Print additional logging output (up to the “info” level)

debug: bool

Print additional logging output (up to the “debug” level).

Passing this flag (or env variable) with no value will enable debug output globally. You may instead pass the name of a specific target function and then only during analysis of that function the debug output is enabled.

trace: booldebug_target: Option<String>result_path: PathBuf

Where to write the resulting GraphLocation (defaults to flow-graph.json)

relaxed: bool

Emit warnings instead of aborting the analysis on sanity checks

This is now the default behavior and this flag is deprecated. Use --strict to turn off this behavior.

strict: bool

Emit errors instead of warnings for potential soundness risks

target: Option<String>

Run paralegal only on this crate

abort_after_analysis: bool

Abort the compilation after finishing the analysis

attach_to_debugger: Option<Debugger>

Attach to a debugger before running the analyses

anactrl: ClapAnalysisCtrl

Additional arguments that control the flow analysis specifically

marker_control: MarkerControl

Additional arguments which control marker assignment and discovery

dump: ParseableDumpArgs

Additional arguments that control debug args specifically

cargo_args: Vec<String>

Pass through for additional cargo arguments (like –features)