Module paralegal_flow::rust

Exposes the rustc external crates (this mod is just to tidy things up).

Re-exports

pub extern crate rustc_abi;

pub extern crate rustc_arena;

pub extern crate rustc_ast;

pub extern crate rustc_borrowck;

pub extern crate rustc_data_structures;

pub extern crate rustc_driver;

pub extern crate rustc_hir;

pub extern crate rustc_interface;

pub extern crate rustc_middle;

pub extern crate rustc_mir_dataflow;

pub extern crate rustc_query_system;

pub extern crate rustc_serialize;

pub extern crate rustc_span;

pub extern crate rustc_target;

pub extern crate rustc_type_ir;

pub use super::rustc_index;

pub use rustc_ast as ast;

pub use rustc_hir as hir;

Modules

mir

sty

ty

Defines how the compiler represents types internally.

Structs

BodyId

DefId

A DefId identifies a particular definition, by combining a crate index and a def index.

LocalDefId

A LocalDefId is equivalent to a DefId with krate == LOCAL_CRATE. Since we encode this information in the type, we can ensure at compile time that no DefIds from upstream crates get thrown into the mix. There are quite a few cases where we know that only DefIds from the local crate are expected; a DefId from a different crate would signify a bug somewhere. This is when LocalDefId comes in handy.

Location

Location represents the position of the start of the statement; or, if statement_index equals the number of statements, then the start of the terminator.

TyCtxt

The central data structure of the compiler. It stores references to the various arenas and also houses the results of the various compiler queries that have been performed. See the rustc dev guide for more details.

Type Aliases

DepGraph