Struct paralegal_flow::rust::ty::TypeckResults

pub struct TypeckResults<'tcx> {

    pub hir_owner: OwnerId,
    type_dependent_defs: UnordMap<ItemLocalId, Result<(DefKind, DefId), ErrorGuaranteed>>,
    field_indices: UnordMap<ItemLocalId, FieldIdx>,
    node_types: UnordMap<ItemLocalId, Ty<'tcx>>,
    node_args: UnordMap<ItemLocalId, &'tcx List<GenericArg<'tcx>>>,
    user_provided_types: UnordMap<ItemLocalId, Canonical<'tcx, UserType<'tcx>>>,
    pub user_provided_sigs: UnordMap<LocalDefId, Canonical<'tcx, Binder<'tcx, FnSig<'tcx>>>>,
    adjustments: UnordMap<ItemLocalId, Vec<Adjustment<'tcx>, Global>>,
    pat_binding_modes: UnordMap<ItemLocalId, BindingMode>,
    pat_adjustments: UnordMap<ItemLocalId, Vec<Ty<'tcx>, Global>>,
    closure_kind_origins: UnordMap<ItemLocalId, (Span, Place<'tcx>)>,
    liberated_fn_sigs: UnordMap<ItemLocalId, FnSig<'tcx>>,
    fru_field_types: UnordMap<ItemLocalId, Vec<Ty<'tcx>, Global>>,
    coercion_casts: UnordSet<ItemLocalId>,
    pub used_trait_imports: UnordSet<LocalDefId>,
    pub tainted_by_errors: Option<ErrorGuaranteed>,
    pub concrete_opaque_types: IndexMap<OpaqueTypeKey<'tcx>, OpaqueHiddenType<'tcx>, BuildHasherDefault<FxHasher>>,
    pub closure_min_captures: UnordMap<LocalDefId, IndexMap<HirId, Vec<CapturedPlace<'tcx>, Global>, BuildHasherDefault<FxHasher>>>,
    pub closure_fake_reads: UnordMap<LocalDefId, Vec<(Place<'tcx>, FakeReadCause, HirId), Global>>,
    pub rvalue_scopes: RvalueScopes,
    pub generator_interior_types: Binder<'tcx, Vec<GeneratorInteriorTypeCause<'tcx>, Global>>,
    pub generator_interior_predicates: UnordMap<LocalDefId, Vec<(Predicate<'tcx>, ObligationCause<'tcx>), Global>>,
    pub treat_byte_string_as_slice: UnordSet<ItemLocalId>,
    pub closure_size_eval: UnordMap<LocalDefId, ClosureSizeProfileData<'tcx>>,
    offset_of_data: UnordMap<ItemLocalId, (Ty<'tcx>, Vec<FieldIdx, Global>)>,

}

Fields
hir_owner: OwnerId

The HirId::owner all ItemLocalIds in this table are relative to.

type_dependent_defs: UnordMap<ItemLocalId, Result<(DefKind, DefId), ErrorGuaranteed>>field_indices: UnordMap<ItemLocalId, FieldIdx>node_types: UnordMap<ItemLocalId, Ty<'tcx>>node_args: UnordMap<ItemLocalId, &'tcx List<GenericArg<'tcx>>>user_provided_types: UnordMap<ItemLocalId, Canonical<'tcx, UserType<'tcx>>>user_provided_sigs: UnordMap<LocalDefId, Canonical<'tcx, Binder<'tcx, FnSig<'tcx>>>>

Stores the canonicalized types provided by the user. See also AscribeUserType statement in MIR.

adjustments: UnordMap<ItemLocalId, Vec<Adjustment<'tcx>, Global>>pat_binding_modes: UnordMap<ItemLocalId, BindingMode>pat_adjustments: UnordMap<ItemLocalId, Vec<Ty<'tcx>, Global>>closure_kind_origins: UnordMap<ItemLocalId, (Span, Place<'tcx>)>liberated_fn_sigs: UnordMap<ItemLocalId, FnSig<'tcx>>fru_field_types: UnordMap<ItemLocalId, Vec<Ty<'tcx>, Global>>coercion_casts: UnordSet<ItemLocalId>used_trait_imports: UnordSet<LocalDefId>

Set of trait imports actually used in the method resolution. This is used for warning unused imports. During type checking, this Lrc should not be cloned: it must have a ref-count of 1 so that we can insert things into the set mutably.

tainted_by_errors: Option<ErrorGuaranteed>

If any errors occurred while type-checking this body, this field will be set to Some(ErrorGuaranteed).

concrete_opaque_types: IndexMap<OpaqueTypeKey<'tcx>, OpaqueHiddenType<'tcx>, BuildHasherDefault<FxHasher>>

All the opaque types that have hidden types set by this function. We also store the type here, so that the compiler can use it as a hint for figuring out hidden types, even if they are only set in dead code (which doesn’t show up in MIR).

closure_min_captures: UnordMap<LocalDefId, IndexMap<HirId, Vec<CapturedPlace<'tcx>, Global>, BuildHasherDefault<FxHasher>>>

Tracks the minimum captures required for a closure; see MinCaptureInformationMap for more details.

closure_fake_reads: UnordMap<LocalDefId, Vec<(Place<'tcx>, FakeReadCause, HirId), Global>>

Tracks the fake reads required for a closure and the reason for the fake read. When performing pattern matching for closures, there are times we don’t end up reading places that are mentioned in a closure (because of _ patterns). However, to ensure the places are initialized, we introduce fake reads. Consider these two examples:

let x: u8;
let c = || match x { _ => () };

In this example, we don’t need to actually read/borrow x in c, and so we don’t want to capture it. However, we do still want an error here, because x should have to be initialized at the point where c is created. Therefore, we add a “fake read” instead.

let c = || {
    let (t1, t2) = t;
}

In the second example, we capture the disjoint fields of t (t.0 & t.1), but we never capture t. This becomes an issue when we build MIR as we require information on t in order to create place t.0 and t.1. We can solve this issue by fake reading t.

rvalue_scopes: RvalueScopes

Tracks the rvalue scoping rules which defines finer scoping for rvalue expressions by applying extended parameter rules. Details may be find in rustc_hir_analysis::check::rvalue_scopes.

generator_interior_types: Binder<'tcx, Vec<GeneratorInteriorTypeCause<'tcx>, Global>>

Stores the type, expression, span and optional scope span of all types that are live across the yield of this generator (if a generator).

generator_interior_predicates: UnordMap<LocalDefId, Vec<(Predicate<'tcx>, ObligationCause<'tcx>), Global>>

Stores the predicates that apply on generator witness types. formatting modified file tests/ui/generator/retain-resume-ref.rs

treat_byte_string_as_slice: UnordSet<ItemLocalId>

We sometimes treat byte string literals (which are of type &[u8; N]) as &[u8], depending on the pattern in which they are used. This hashset records all instances where we behave like this to allow const_to_pat to reliably handle this situation.

closure_size_eval: UnordMap<LocalDefId, ClosureSizeProfileData<'tcx>>

Contains the data for evaluating the effect of feature capture_disjoint_fields on closure size.

offset_of_data: UnordMap<ItemLocalId, (Ty<'tcx>, Vec<FieldIdx, Global>)>