Module paralegal_flow::serializers

serde serializers, most used for JSON debugging output in crate::dbg.

The proxy structs are foreign serializers for their non-proxy counterparts, see https://serde.rs/remote-derive.html for more information. As a naming convention <name>Proxy is used to (de)serialize <name> e.g. [BasicBlockProxy] (de)serializes a mir::BasicBlock.

Be aware that in some cases serialization is not bidirectional (usually if there is a lifetime parameter in the serialized type). For instance [GlobalLocation] can be serialized, but only a [RawGlobalLocation] can be deserialized.

Some types (such as mir::Body) first have to be explicitly transformed into the respective proxy type. In the case of mir::Body this can be done with BodyProxy::from_body_with_normalize

Structs

Bodies

A serializable version of mir::Bodys, mapped to their [hir::BodyId] so that you can resolve the body belonging to a global location (see [IsGlobalLocation::function]).

BodyIdProxy2

This exists because of serde’s restrictions on how you derive serializers. [BodyIdProxy] can be used to serialize a BodyId but if the BodyId is used as e.g. a key in a map or in a vector it does not dispatch to the remote impl on [BodyIdProxy]. Implementing the serializers for the map or vector by hand is annoying so instead you can map over the datastructure, wrap each BodyId in this proxy type and then dispatch to the serialize impl for the reconstructed data structure.

BodyProxy

A serializable version of a mir::Body.

InstructionProxy

Functions

iter_stmts 🔒