paralegal_flow
Struct MarkerControl

pub struct MarkerControl {
    external_annotations: Option<PathBuf>,
}

Fields
external_annotations: Option<PathBuf>

A JSON file from which to load additional annotations. Whereas normally annotation can only be placed on crate-local items, these can also be placed on third party items, such as functions from the stdlib.

The file is expected to contain a HashMap<Identifier, (Vec<Annotation>, ObjectType)>, which is the same type as annotations field from the ProgramDescription struct. It uses the serde derived serializer. An example for the format can be generated by running paralegal-flow with dump_serialized_flow_graph.

Implementations

impl MarkerControl

pub fn external_annotations(&self) -> Option<&Path>