Module paralegal_flow::utils

Utility functions, general purpose structs and extension traits

Modules

print 🔒

resolve

Structs

DisplayViaDebug

PlaceVisitor

A struct that can be used to apply a FnMut to every Place in a MIR object via the visit::Visitor trait. Usually used to accumulate information about the places.

Print

RePlacer

A struct that can be used to apply a FnMut to every Place in a MIR object via the MutVisitor trait. Crucial difference to PlaceVisitor is that this function can alter the place itself.

RecursionBreakingCache

This code is adapted from [flowistry::cached::Cache] but with a recursion breaking mechanism. This alters the Self::get method signature to return an Option of a reference. In particular the method will return None if it is called with the same key while computing a construction function for that key.

ShortHash

Create a hash for this object that is no longer than six hex digits

TinyBitSet

A bit-set implemented with a u16, supporting domains up to 16 elements.

Enums

AsFnAndArgsErr

BodyResolutionError

FnResolution

FunctionKind 🔒

This exists to distinguish different types of functions, which is necessary because depending on the type of function, the method of requesting its signature from TyCtxt differs.

Overlap

StmtAtErr

Traits

AsFnAndArgs

Extension trait for function calls (e.g. mir::TerminatorKind and mir::Terminator) that lets you decompose the call into a convenient (function definition, arguments, return place) tuple in such cases when

DfppBodyExt

FnResolutionExt

GenericArgExt

IntoBodyId

IntoDefId

Brother to IntoLocalDefId, converts the id type to a DefId using TyCtxt

IntoHirId

IntoLocalDefId

A trait for types that can be converted into a [mir::LocalDefId] via TyCtxt.

MetaItemMatch

This is meant as an extension trait for ast::Attribute. The main method of interest is match_extract, matches_path is interesting if you want to check if this attribute has the path of interest but you do not care about its payload.

NodeExt

Extension trait for [hir::Node]. This lets up implement methods on [hir::Node]. Self should only ever be instantiated as [hir::Node].

PlaceExt

Extension trait for Places so we can implement methods on them. Self is only ever supposed to be instantiated as Place.

ProjectionElemExt

Spanned

TyCtxtExt

Extension trait for TyCtxt

TyExt

Extension trait for ty::Ty. This lets us implement methods on ty::Ty. Self is only ever supposed to be instantiated as ty::Ty.

Functions

body_name_pls

Get the name of the function for this body as an Ident. This handles such cases correctly where the function in question has no proper name, as is the case for closures.

dump_file_pls

Create a file for dumping an ext kind of output for id. The name of the resulting file avoids clashes but is also descriptive (uses the resolved name of id).

extract_places

Old version of places_read, should be considered deprecated.

handle_aggregate_assign 🔒

identifier_for_item

Get a reasonable, but not guaranteed unique name for this item

outfile_pls

Give me this file as writable (possibly creating or overwriting it).

places_read

Return all places mentioned at this location that are read. Which means that if a Place is not read but assigned (e.g. the return place of a function call), it will not be in the result set.

read_places_with_provenance

Return the places that are read in this statements and possible ref/deref un-layerings of those places.

test_generics_normalization 🔒

Try and normalize the provided generics.

time

unique_and_terse_body_name_pls

Gives a string name for this i that is free of name clashes, as it includes a hash of the id.

unique_identifier_for_item

Creates an Identifier for this HirId

with_temporary_logging_level

write_sep

Write all elements from it into the formatter fmt using f, separating them with sep

Type Aliases

SimplifiedArguments

A simplified version of the argument list that is stored in a TerminatorKind::Call.