paralegal_flow::utils
Module resolve

Enums

Res

ResolutionError

SearchSpace

Functions

as_primitive_ty 🔒

def_path_res

Main algorithm lifted from clippy_utils. I’ve made additions so this can handle some qualified paths.

expect_resolve_string_to_def_id

A small helper wrapper around def_path_res that represents a common way that def_path_res is used. In the case of errors they are reported to the user and None is returned so the caller has the option of making progress before exiting.

find_crates 🔒

find_primitive_impls 🔒

item_child_by_name 🔒

local_item_children_by_name 🔒

non_local_item_children_by_name 🔒

resolve_ty 🔒

Type Aliases

Result