paralegal_flow::utils
Trait InstanceExt

pub trait InstanceExt<'tcx> {
    // Required method
    fn sig(self, tcx: TyCtxt<'tcx>) -> Result<FnSig<'tcx>, ErrorGuaranteed>;
}

Required Methods

fn sig(self, tcx: TyCtxt<'tcx>) -> Result<FnSig<'tcx>, ErrorGuaranteed>

Get the most precise type signature we can for this function, erase any regions and discharge binders.

Returns an error if it was impossible to get any signature.

Emits warnings if a precise signature could not be obtained or there were type variables not instantiated.