Trait paralegal_flow::utils::TyCtxtExt

pub trait TyCtxtExt<'tcx> {
    // Required methods
    fn body_for_def_id(
        self,
        local_def_id: LocalDefId
    ) -> Result<&'tcx BodyWithBorrowckFacts<'tcx>, BodyResolutionError>;
    fn body_for_def_id_default_policy(
        self,
        local_def_id: LocalDefId
    ) -> Option<&'tcx BodyWithBorrowckFacts<'tcx>>;
}

Extension trait for TyCtxt

Required Methods

fn body_for_def_id( self, local_def_id: LocalDefId ) -> Result<&'tcx BodyWithBorrowckFacts<'tcx>, BodyResolutionError>

Resolve this DefId to a body. Returns BodyWithBorrowckFacts, because it internally uses flowistry’s body resolution (rustc_utils::mir::borrowck_facts::get_body_with_borrowck_facts) which memoizes its results so this is actually a cheap query.

Returns None if the id does not refer to a function or if its body is unavailable.

fn body_for_def_id_default_policy( self, local_def_id: LocalDefId ) -> Option<&'tcx BodyWithBorrowckFacts<'tcx>>

Essentially the same as Self::body_for_def_id but handles errors according to our default policy which is as follows:

BodyResolutionError::NotAFunction: Hard error (panic). We consider this an ICE because calling this method on a non-function DefId could indicate errors elsewhere in the compiler.

BodyResolutionError::External: Silent because otherwise we would spam warnings.

BodyResolutionError::IsTraitAssocFn: Warning emitted, because this is probably caused by dyn, which we can’t resolve even if it’s crate-local and that might be surprising.