paralegal_flow
Module ann

Modules

db

Central repository for information about markers (and annotations).

parse

nom-based parser-combinators for parsing the token stream in Rust Attributes.

Structs

DumpingVisitor 🔒

ExceptionAnnotation

MarkerAnnotation

A marker annotation and its refinements.

MarkerRefinement

Refinements in the marker targeting. The default (no refinement provided) is on_argument == vec![] and on_return == false, which is also what is returned from Self::empty.

Markers 🔒

These are pseudo-constants that are used to match annotations. In theory they never change but they use [Symbol] inside, which is only valid so long as the rustc SESSION_GLOBALS are live so we need to calculate them afresh in new.

VisitFilter 🔒

Enums

Annotation

Types of annotations we support.

MarkerRefinementKind

Disaggregated version of MarkerRefinement. Can be added to an existing refinement MarkerRefinement::merge_kind.

Constants

MARKER_META_EXT 🔒

Functions

const_false 🔒

dump_markers

A visit over the HIR that collects all the marker annotations we can find and dumps them out

Type Aliases

MarkerMeta 🔒

VerificationHash