paralegal_flow::ann
Module parse

nom-based parser-combinators for parsing the token stream in Rust Attributes.

Usually used in a closure handed to MetaItemMatch::match_extract.

The benefit of using a combinator library such as nom is not just that it gives us boundaries for parsers that lets us (re)combine them, but also that we get features that are annoying to implement (such as backtracking) for free.

Re-exports

pub extern crate nom;

Structs

DisplaySpan ğŸ”’

I

Just a newtype-wrapper for CursorRef so we can implement traits on it (specifically nom::InputLength).

NotAnIdentifierErr ğŸ”’

NotIdentifierErr ğŸ”’

RefinementOnNonFunctionErr ğŸ”’

StructuralError ğŸ”’

Symbols

UnusedInputErr ğŸ”’

WrongTokenKindErr ğŸ”’

Enums

Structure ğŸ”’

Functions

ann_match_fn ğŸ”’

Parser for a [LabelAnnotation]

apply_parser_in_delimited_annotation ğŸ”’

arguments

Parse the arguments refinement.

assert_identifier

Expect the next token to be an identifier with the value s

assert_token

Expect the next token to have the token kind k.

delimited

Parse a TokenTree::Delimited with the delimiter character delim, applying the subparser p to the tokens in between the delimiters and return the result of the subparser.

dict

identifier

Parse an identifier. Identifiers in annotations are similar to identifiers in rust in general, e.g. strings or word character, numbers and underscores.

integer

Parse an integer literal and return the integer.

integer_list

Parse bracket-delimited, comma-separated integers, e.g. [1,2,3].

lit

Parse a TokenKind::Literal if it has a specific LitKind and return the payload of the literal.

match_exception ğŸ”’

Parser for an ExceptionAnnotation

one ğŸ”’

Parse any one token, returning the token.

one_token

Parse a single token that is not a subtree and return the token.

otype_ann_match ğŸ”’

Parser for the payload of the #[paralegal_flow::output_type(...)] annotation.

refinements_parser ğŸ”’

A parser for annotation refinements.

Type Aliases

Integer ğŸ”’

R ğŸ”’