paralegal_flow
Module utils

Utility functions, general purpose structs and extension traits

Modules

print 🔒

resolve

Structs

DisplayViaDebug

PlaceVisitor

A struct that can be used to apply a FnMut to every Place in a MIR object via the visit::Visitor trait. Usually used to accumulate information about the places.

Print

ShortHash

Create a hash for this object that is no longer than six hex digits

TinyBitSet

A bit-set implemented with a u16, supporting domains up to 16 elements.

Enums

AsFnAndArgsErr

FunctionKind

This exists to distinguish different types of functions, which is necessary because depending on the type of function, the method of requesting its signature from TyCtxt differs.

Overlap

StmtAtErr

Traits

AsFnAndArgs

Extension trait for function calls (e.g. mir::TerminatorKind and mir::Terminator) that lets you decompose the call into a convenient (function definition, arguments, return place) tuple in such cases when

DfppBodyExt

GenericArgExt

InstanceExt

IntoDefId

Brother to IntoLocalDefId, converts the id type to a DefId using TyCtxt

IntoLocalDefId

A trait for types that can be converted into a [mir::LocalDefId] via TyCtxt.

MetaItemMatch

This is meant as an extension trait for ast::Attribute. The main method of interest is match_extract, matches_path is interesting if you want to check if this attribute has the path of interest but you do not care about its payload.

NodeExt

Extension trait for [hir::Node]. This lets up implement methods on [hir::Node]. Self should only ever be instantiated as [hir::Node].

PlaceExt

Extension trait for Places so we can implement methods on them. Self is only ever supposed to be instantiated as Place.

ProjectionElemExt

Spanned

TyExt

Extension trait for ty::Ty. This lets us implement methods on ty::Ty. Self is only ever supposed to be instantiated as ty::Ty.

Functions

body_span

This function exists to deal with #[tracing::instrument]. In that case, sadly, the Span value attached to a body directly refers only to the #[tracing::instrument] macro call. This function instead reconstitutes the span from the collection of spans on each statement.

func_of_term

identifier_for_item

Get a reasonable, but not guaranteed unique name for this item

is_virtual

Returns whether this method is expected to have a body we can analyze.

map_either

test_generics_normalization 🔒

Try and normalize the provided generics.

time

ty_of_const

type_for_constructor

with_temporary_logging_level

write_sep

Write all elements from it into the formatter fmt using f, separating them with sep

Type Aliases

SimplifiedArguments

A simplified version of the argument list that is stored in a TerminatorKind::Call.